+++ client_test.go
@@ -21,6 +21,12 @@ func testClient() Client {
 	return client
 }
 
+func authenticatedTestClient() Client {
+	client := testClient()
+	client.Token = "ABC"
+	return client
+}
+
 // ErrReader implements the io.Reader interface and fails on Read.
 type ErrReader struct{}
 
@@ -58,7 +64,7 @@ func TestClientLogin(t *testing.T) {
 // TestClientGet tests the Client::Get method.
 func TestClientGet(t *testing.T) {
 	defer gock.Off()
-	client := testClient()
+	client := authenticatedTestClient()
 	var err error
 
 	// Success
@@ -91,7 +97,7 @@ func TestClientGet(t *testing.T) {
 // TestClientDeleteDn tests the Client::Delete method.
 func TestClientDelete(t *testing.T) {
 	defer gock.Off()
-	client := testClient()
+	client := authenticatedTestClient()
 
 	// Success
 	gock.New(testURL).
@@ -111,7 +117,7 @@ func TestClientDelete(t *testing.T) {
 // TestClientPost tests the Client::Post method.
 func TestClientPost(t *testing.T) {
 	defer gock.Off()
-	client := testClient()
+	client := authenticatedTestClient()
 
 	var err error
 
@@ -145,7 +151,7 @@ func TestClientPost(t *testing.T) {
 // TestClientPost tests the Client::Post method.
 func TestClientPut(t *testing.T) {
 	defer gock.Off()
-	client := testClient()
+	client := authenticatedTestClient()
 
 	var err error