+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.1.2 (unreleased)
+
+- Return an error if authentication fails
+- Add mutex to synchronize authentications
+
 ## 0.1.1
 
 - Only retry on 408 and 5xx HTTP response status codes